+++ cmake/ThirdPartyHDF5.cmake
@@ -6,24 +6,39 @@
 #
 ########################################################################
 
+# Default HDF5 on for non-Windows platforms
+IF(WIN32)
+    SET(NEKTAR_DEFAULT_HDF5 OFF)
+ELSE()
+    SET(NEKTAR_DEFAULT_HDF5 ON)
+ENDIF()
+
 OPTION(NEKTAR_USE_HDF5
-    "Enable HDF5 I/O support." OFF)
+    "Enable HDF5 I/O support." ${NEKTAR_DEFAULT_HDF5})
+
+UNSET(NEKTAR_DEFAULT_HDF5)
 
 IF (NEKTAR_USE_HDF5)
-    IF (NOT NEKTAR_USE_MPI)
-        MESSAGE(FATAL_ERROR "HDF5 requires Nektar++ to be configured with "
-                "NEKTAR_USE_MPI for MPI support.")
+    # Try to find parallel system HDF5 first.
+    IF (NEKTAR_USE_MPI)
+        SET(HDF5_PREFER_PARALLEL ON)
     ENDIF()
 
-    # Try to find parallel system HDF5 first.
-    SET(HDF5_PREFER_PARALLEL ON)
     FIND_PACKAGE(HDF5 QUIET)
 
-    IF (HDF5_FOUND AND NOT HDF5_IS_PARALLEL)
-        MESSAGE(STATUS "Non-parallel system HDF5 detected: will build instead.")
-        SET(BUILD_HDF5 ON)
-    ELSEIF(HDF5_FOUND)
+    IF (HDF5_IS_PARALLEL)
+        ADD_DEFINITIONS(-DNEKTAR_HDF5_PARALLEL)
+    ENDIF()
+
+    IF (HDF5_FOUND)
         SET(BUILD_HDF5 OFF)
+
+        IF (NOT HDF5_IS_PARALLEL AND NEKTAR_USE_MPI)
+            MESSAGE(WARNING "Using non-parallel HDF5 with MPI enabled: possible performance issues.")
+        ELSEIF (NOT NEKTAR_USE_MPI)
+            # Parallel built HDF5 will not compile against non-MPI enabled code.
+            SET(BUILD_HDF5 ON)
+        ENDIF()
     ELSE()
         SET(BUILD_HDF5 ON)
     ENDIF()
@@ -33,8 +48,26 @@ IF (NEKTAR_USE_HDF5)
         "NEKTAR_USE_HDF5" OFF)
 
     IF(THIRDPARTY_BUILD_HDF5)
+        IF (NEKTAR_USE_MPI)
+            SET(HDF5_MPI_CONFIG -DHDF5_ENABLE_PARALLEL=ON)
+        ELSE()
+            SET(HDF5_MPI_CONFIG -DHDF5_ENABLE_PARALLEL=OFF)
+        ENDIF()
+
         INCLUDE(ExternalProject)
-        THIRDPARTY_LIBRARY(HDF5_LIBRARIES SHARED hdf5
+
+        # In debug mode HDF5 is built either as hdf5_D or hdf5_debug
+        IF (CMAKE_BUILD_TYPE STREQUAL "Debug")
+            IF (WIN32)
+                set(HDF5_LIB_NAME hdf5_D)
+            ELSE()
+                set(HDF5_LIB_NAME hdf5_debug)
+            ENDIF()
+        ELSE()
+            SET(HDF5_LIB_NAME hdf5)
+        ENDIF()
+
+        THIRDPARTY_LIBRARY(HDF5_LIBRARIES SHARED ${HDF5_LIB_NAME}
             DESCRIPTION "HDF5 library")
 
         EXTERNALPROJECT_ADD(
@@ -56,13 +89,25 @@ IF (NEKTAR_USE_HDF5)
                 -DCMAKE_C_FLAGS="-w"
                 -DCMAKE_CXX_FLAGS="-w"
                 -DCMAKE_INSTALL_PREFIX:PATH=${TPDIST}
-                -DDEFAULT_API_VERSION=v110
-                -DHDF5_ENABLE_PARALLEL=ON
+		-DDEFAULT_API_VERSION=v110
+                ${HDF5_MPI_CONFIG}
                 -DHDF5_BUILD_CPP_LIB=OFF
                 -DBUILD_TESTING=OFF
                 -DHDF5_BUILD_TOOLS=OFF
                 ${TPSRC}/hdf5-1.12.3
             )
+
+        # Add definition to enable dllexport flag.
+        IF(WIN32)
+            ADD_DEFINITIONS(-DH5_BUILT_AS_DYNAMIC_LIB)
+
+            # Make sure dlls are installed
+            INSTALL(CODE "FILE(GLOB hdf5dlls \"${TPDIST}/bin/hdf5*.dll\")
+                IF (NOT hdf5dlls STREQUAL \"\")
+                    FILE(INSTALL \${hdf5dlls} DESTINATION \${CMAKE_INSTALL_PREFIX}/${NEKTAR_BIN_DIR})
+                ENDIF()
+            ")
+        ENDIF()
 
         SET(HDF5_INCLUDE_DIRS ${TPDIST}/include CACHE FILEPATH
             "HDF5 include directory" FORCE)